reserve one train file
