+++ example/distill/mnist_distill/train_with_fleet.py
@@ -1,4 +1,4 @@
-#   Copyright (c) 2018 PaddlePaddle Authors. All Rights Reserved.
+#   Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,28 +13,36 @@
 # limitations under the License.
 
 from __future__ import print_function
-
 import os
+
+if os.environ.get('PADDLE_TRAINER_ENDPOINTS') is None:
+    os.environ['PADDLE_TRAINER_ENDPOINTS'] = '127.0.0.1:0'
+
+from paddle_edl.distill.distill_reader import DistillReader
 import argparse
+import ast
 from PIL import Image
 import numpy
 import paddle
 import paddle.fluid as fluid
+from paddle.fluid.incubate.fleet.collective import fleet
+from paddle.fluid.incubate.fleet.base import role_maker
 
 
 def parse_args():
     parser = argparse.ArgumentParser("mnist")
     parser.add_argument(
-        '--enable_ce',
-        action='store_true',
-        help="If set, run the task with continuous evaluation logs.")
-    parser.add_argument(
         '--use_gpu',
-        type=bool,
-        default=False,
-        help="Whether to use GPU or not.")
+        type=ast.literal_eval,
+        default=True,
+        help="Whether to use GPU or not. 'True' or 'False'")
     parser.add_argument(
         '--num_epochs', type=int, default=5, help="number of epochs.")
+    parser.add_argument(
+        '--use_distill_service',
+        default=False,
+        type=ast.literal_eval,
+        help="Whether to use distill service train. 'True' or 'False'")
     args = parser.parse_args()
     return args
 
@@ -87,20 +95,13 @@ def train(nn_type,
     startup_program = fluid.default_startup_program()
     main_program = fluid.default_main_program()
 
-    if args.enable_ce:
-        train_reader = paddle.batch(
-            paddle.dataset.mnist.train(), batch_size=BATCH_SIZE)
-        test_reader = paddle.batch(
-            paddle.dataset.mnist.test(), batch_size=BATCH_SIZE)
-        startup_program.random_seed = 90
-        main_program.random_seed = 90
-    else:
-        train_reader = paddle.batch(
-            paddle.reader.shuffle(
-                paddle.dataset.mnist.train(), buf_size=500),
-            batch_size=BATCH_SIZE)
-        test_reader = paddle.batch(
-            paddle.dataset.mnist.test(), batch_size=BATCH_SIZE)
+    train_reader = paddle.batch(
+        paddle.reader.shuffle(
+            paddle.dataset.mnist.train(), buf_size=500),
+        batch_size=BATCH_SIZE)
+
+    test_reader = paddle.batch(
+        paddle.dataset.mnist.test(), batch_size=BATCH_SIZE)
 
     img = fluid.data(name='img', shape=[None, 1, 28, 28], dtype='float32')
     label = fluid.data(name='label', shape=[None, 1], dtype='int64')
@@ -115,15 +116,45 @@ def train(nn_type,
     prediction, avg_loss, acc = net_conf(img, label)
 
     test_program = main_program.clone(for_test=True)
-    optimizer = fluid.optimizer.Adam(learning_rate=0.001)
-    optimizer.minimize(avg_loss)
 
-    def train_test(train_test_program, train_test_feed, train_test_reader):
+    inputs = [img, label]
+    test_inputs = [img, label]
+
+    if args.use_distill_service:
+        dr = DistillReader(ins=['img', 'label'], predicts=['prediction'])
+        train_reader = dr.set_sample_list_generator(train_reader)
+
+        soft_label = fluid.data(
+            name='soft_label', shape=[None, 10], dtype='float32')
+        inputs.append(soft_label)
+        distill_loss = fluid.layers.cross_entropy(
+            input=prediction, label=soft_label, soft_label=True)
+        #distill_loss = fluid.layers.mse_loss(input=prediction, label=soft_label)
+        distill_loss = fluid.layers.mean(distill_loss)
+        #loss = 0.3 * avg_loss + distill_loss
+        loss = distill_loss
+    else:
+        loss = avg_loss
+
+    role = role_maker.PaddleCloudRoleMaker(is_collective=True)
+    fleet.init(role)
+    train_rank = fleet.worker_index()
+    train_nranks = fleet.worker_num()
+
+    optimizer = fluid.optimizer.Momentum(learning_rate=0.001, momentum=0.9)
+    if train_nranks != 1:
+        optimizer = fleet.distributed_optimizer(optimizer)
+    optimizer.minimize(loss)
+
+    main_program = fleet.main_program if train_nranks != 1 else main_program
+    gpu_id = int(os.getenv("FLAGS_selected_gpus", "0"))
+
+    def train_test(train_test_program, train_test_reader):
         acc_set = []
         avg_loss_set = []
         for test_data in train_test_reader():
             acc_np, avg_loss_np = exe.run(program=train_test_program,
-                                          feed=train_test_feed.feed(test_data),
+                                          feed=test_data,
                                           fetch_list=[acc, avg_loss])
             acc_set.append(float(acc_np))
             avg_loss_set.append(float(avg_loss_np))
@@ -132,55 +163,59 @@ def train(nn_type,
         avg_loss_val_mean = numpy.array(avg_loss_set).mean()
         return avg_loss_val_mean, acc_val_mean
 
-    place = fluid.CUDAPlace(0) if use_cuda else fluid.CPUPlace()
+    place = fluid.CUDAPlace(gpu_id) if use_cuda else fluid.CPUPlace()
+    reader_places = fluid.cuda_places() if use_cuda else fluid.CPUPlace()
+
+    py_train_reader = fluid.io.DataLoader.from_generator(
+        feed_list=inputs, capacity=64)
+    py_train_reader.set_sample_list_generator(train_reader, reader_places)
+    py_test_reader = fluid.io.DataLoader.from_generator(
+        feed_list=test_inputs, capacity=64)
+    py_test_reader.set_sample_list_generator(test_reader, reader_places)
 
     exe = fluid.Executor(place)
-
-    feeder = fluid.DataFeeder(feed_list=[img, label], place=place)
     exe.run(startup_program)
     epochs = [epoch_id for epoch_id in range(PASS_NUM)]
 
     lists = []
     step = 0
     for epoch_id in epochs:
-        for step_id, data in enumerate(train_reader()):
-            metrics = exe.run(main_program,
-                              feed=feeder.feed(data),
-                              fetch_list=[avg_loss, acc])
+        for step_id, data in enumerate(py_train_reader()):
+            metrics = exe.run(main_program, feed=data, fetch_list=[loss, acc])
             if step % 100 == 0:
-                print("Pass %d, Epoch %d, Cost %f" % (step, epoch_id,
-                                                      metrics[0]))
+                print("Pass {}, Epoch {}, Cost {}".format(step, epoch_id,
+                                                          metrics[0]))
             step += 1
-        # test for epoch
-        avg_loss_val, acc_val = train_test(
-            train_test_program=test_program,
-            train_test_reader=test_reader,
-            train_test_feed=feeder)
 
-        print("Test with Epoch %d, avg_cost: %s, acc: %s" %
-              (epoch_id, avg_loss_val, acc_val))
-        lists.append((epoch_id, avg_loss_val, acc_val))
+        if train_rank == 0:
+            # test for epoch
+            avg_loss_val, acc_val = train_test(
+                train_test_program=test_program,
+                train_test_reader=py_test_reader)
 
-        if save_dirname is not None:
-            fluid.io.save_inference_model(
-                save_dirname, ["img"], [prediction],
-                exe,
-                model_filename=model_filename,
-                params_filename=params_filename)
+            print("Test with Epoch %d, avg_cost: %s, acc: %s" %
+                  (epoch_id, avg_loss_val, acc_val))
+            lists.append((epoch_id, avg_loss_val, acc_val))
+            if save_dirname is not None:
+                fluid.io.save_inference_model(
+                    save_dirname, ["img"], [prediction],
+                    exe,
+                    model_filename=model_filename,
+                    params_filename=params_filename)
 
-    import paddle_serving_client.io as serving_io
-    serving_io.save_model("mnist_model", "mnist_client_conf", {"img": img},
-                          {"prediction": prediction}, test_program)
+    if train_rank == 0:
+        if args.save_serving_model:
+            import paddle_serving_client.io as serving_io
+            serving_io.save_model("mnist_model", "mnist_client_conf",
+                                  {img.name: img},
+                                  {prediction.name: prediction}, test_program)
+            print('save serving model, feed_names={}, fetch_names={}'.format(
+                [img.name], [prediction.name]))
 
-    if args.enable_ce:
-        print("kpis\ttrain_cost\t%f" % metrics[0])
-        print("kpis\ttest_cost\t%s" % avg_loss_val)
-        print("kpis\ttest_acc\t%s" % acc_val)
-
-    # find the best pass
-    best = sorted(lists, key=lambda list: float(list[1]))[0]
-    print('Best pass is %s, testing Avgcost is %s' % (best[0], best[1]))
-    print('The classification accuracy is %.2f%%' % (float(best[2]) * 100))
+        # find the best pass
+        best = sorted(lists, key=lambda list: float(list[1]))[0]
+        print('Best pass is %s, testing Avgcost is %s' % (best[0], best[1]))
+        print('The classification accuracy is %.2f%%' % (float(best[2]) * 100))
 
 
 def infer(use_cuda,
@@ -190,13 +225,14 @@ def infer(use_cuda,
     if save_dirname is None:
         return
 
-    place = fluid.CUDAPlace(0) if use_cuda else fluid.CPUPlace()
+    gpu_id = int(os.getenv("FLAGS_selected_gpus", "0"))
+    place = fluid.CUDAPlace(gpu_id) if use_cuda else fluid.CPUPlace()
     exe = fluid.Executor(place)
 
     def load_image(file):
         im = Image.open(file).convert('L')
         im = im.resize((28, 28), Image.ANTIALIAS)
-        im = numpy.array(im).reshape(1, 1, 28, 28).astype(numpy.float32)
+        im = numpy.array(im).reshape((1, 1, 28, 28)).astype(numpy.float32)
         im = im / 255.0 * 2.0 - 1.0
         return im
 
@@ -247,6 +283,6 @@ if __name__ == '__main__':
     PASS_NUM = args.num_epochs
     use_cuda = args.use_gpu
     # predict = 'softmax_regression' # uncomment for Softmax
-    # predict = 'multilayer_perceptron' # uncomment for MLP
+    #predict = 'multilayer_perceptron' # uncomment for MLP
     predict = 'convolutional_neural_network'  # uncomment for LeNet5
     main(use_cuda=use_cuda, nn_type=predict)
